projects: move shuffle to utils
