+++ src/utils/shuffle.ts
@@ -1,0 +1,12 @@
+export default function shuffle<Type>(array: Type[]): Type[] {
+  let i = array.length - 1;
+  for (; i > 0; i--) {
+    const j = Math.floor(Math.random() * (i + 1));
+    const temp = array[i];
+    // eslint-disable-next-line no-param-reassign
+    array[i] = array[j];
+    // eslint-disable-next-line no-param-reassign
+    array[j] = temp;
+  }
+  return array;
+}
